allow stratify.cols to be integer

+++ R/ResampleInstance.R
@@ -91,7 +91,7 @@ makeResampleInstance = function(desc, task, size, ...) {
     if (length(i) > 0L)
       stopf("Columns specified for stratification, but not present in task: %s", collapse(stratify.cols[i]))
     index = getTaskData(task, features = stratify.cols, target.extra = FALSE)[stratify.cols]
-    if (any(vlapply(index, is.numeric)))
+    if (any(vlapply(index, is.double)))
       stop("Stratification on numeric variables not possible")
     grp = tapply(seq_row(index), index, simplify = FALSE)
     grp = unname(split(seq_row(index), grp))
